# docs/contracts/openapi-v1.yaml
openapi: 3.0.3
info:
  title: SID3 API
  version: 1.0.0
  description: Drive-backed storage gateway API for SID3 MVP.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Projects
  - name: Integrations
  - name: API Keys
  - name: Buckets
  - name: Objects
  - name: Logs
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a SID3 user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login a SID3 user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags: [Auth]
      summary: Return current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /projects:
    get:
      tags: [Projects]
      summary: List projects owned by the current user
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /integrations/google/authorize:
    post:
      tags: [Integrations]
      summary: Create a Google OAuth authorization URL
      responses:
        '200':
          description: Authorization URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthAuthorizeResponse'
  /integrations/google/callback:
    post:
      tags: [Integrations]
      summary: Complete Google OAuth connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthCallbackRequest'
      responses:
        '200':
          description: Connected integration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderIntegration'
  /integrations:
    get:
      tags: [Integrations]
      summary: List provider integrations
      responses:
        '200':
          description: Integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderIntegration'
  /integrations/{integrationId}/revoke:
    post:
      tags: [Integrations]
      summary: Revoke a provider integration
      parameters:
        - $ref: '#/components/parameters/IntegrationId'
      responses:
        '200':
          description: Revoked integration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderIntegration'
  /projects/{projectId}/api-keys:
    get:
      tags: [API Keys]
      summary: List project API keys
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
    post:
      tags: [API Keys]
      summary: Create a project API key
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: Created API key; secret is returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
  /projects/{projectId}/api-keys/{apiKeyId}:
    delete:
      tags: [API Keys]
      summary: Revoke a project API key
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '204':
          description: Revoked
  /projects/{projectId}/buckets:
    get:
      tags: [Buckets]
      summary: List project buckets
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bucket'
    post:
      tags: [Buckets]
      summary: Create a virtual bucket
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBucketRequest'
      responses:
        '201':
          description: Created bucket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bucket'
  /projects/{projectId}/logs:
    get:
      tags: [Logs]
      summary: List project operation logs
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Operation logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperationLog'
  /buckets/{bucketId}/objects:
    get:
      tags: [Objects]
      security:
        - apiKeyAuth: []
      summary: List objects in a bucket
      parameters:
        - $ref: '#/components/parameters/BucketId'
        - name: prefix
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectListResponse'
    post:
      tags: [Objects]
      security:
        - apiKeyAuth: []
      summary: Upload an object
      parameters:
        - $ref: '#/components/parameters/BucketId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadObjectRequest'
      responses:
        '201':
          description: Uploaded object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageObject'
  /buckets/{bucketId}/objects/{objectId}/download:
    get:
      tags: [Objects]
      security:
        - apiKeyAuth: []
      summary: Download an object
      parameters:
        - $ref: '#/components/parameters/BucketId'
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '200':
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /buckets/{bucketId}/objects/{objectId}:
    delete:
      tags: [Objects]
      security:
        - apiKeyAuth: []
      summary: Delete an object
      parameters:
        - $ref: '#/components/parameters/BucketId'
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '204':
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-SID3-API-Key
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BucketId:
      name: bucketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ObjectId:
      name: objectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ApiKeyId:
      name: apiKeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IntegrationId:
      name: integrationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        name:
          type: string
          minLength: 2
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
    Project:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
    OAuthAuthorizeResponse:
      type: object
      required: [authorizationUrl, stateExpiresAt]
      properties:
        authorizationUrl:
          type: string
          format: uri
        stateExpiresAt:
          type: string
          format: date-time
    OAuthCallbackRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
        state:
          type: string
    ProviderIntegration:
      type: object
      required: [id, provider, status, scopes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [GOOGLE_DRIVE]
        providerAccountEmail:
          type: string
          format: email
          nullable: true
        status:
          type: string
          enum: [CONNECTED, REVOKED, ERROR]
        scopes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
    ApiKey:
      type: object
      required: [id, name, prefix, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ApiKeyCreated:
      type: object
      required: [apiKey, secret]
      properties:
        apiKey:
          $ref: '#/components/schemas/ApiKey'
        secret:
          type: string
          description: Returned once; store securely.
    CreateBucketRequest:
      type: object
      required: [name, providerIntegrationId]
      properties:
        name:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{2,62}$'
        providerIntegrationId:
          type: string
          format: uuid
    Bucket:
      type: object
      required: [id, projectId, name, providerIntegrationId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        providerIntegrationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    UploadObjectRequest:
      type: object
      required: [key, file]
      properties:
        key:
          type: string
          minLength: 1
        file:
          type: string
          format: binary
    StorageObject:
      type: object
      required: [id, bucketId, key, fileName, contentType, sizeBytes, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        bucketId:
          type: string
          format: uuid
        key:
          type: string
        fileName:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        checksumSha256:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, AVAILABLE, DELETING, DELETED, FAILED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ObjectListResponse:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StorageObject'
        nextCursor:
          type: string
          nullable: true
    OperationLog:
      type: object
      required: [id, projectId, operation, status, requestId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        apiKeyId:
          type: string
          format: uuid
          nullable: true
        bucketId:
          type: string
          format: uuid
          nullable: true
        objectId:
          type: string
          format: uuid
          nullable: true
        operation:
          type: string
          enum: [UPLOAD, DOWNLOAD, LIST, DELETE, OAUTH_CONNECT, OAUTH_REVOKE, API_KEY_CREATE, API_KEY_REVOKE]
        status:
          type: string
          enum: [SUCCESS, FAILED]
        provider:
          type: string
          enum: [GOOGLE_DRIVE]
          nullable: true
        errorCode:
          type: string
          nullable: true
        requestId:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string